chore: rename avalonia namespace to ColDogStudios.ColDogLocker.Avalonia
