+++ ColDogLocker.Avalonia/App.axaml
@@ -1,14 +1,14 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             x:Class="ColDogLocker.Gui.Avalonia.App"
-             xmlns:local="using:ColDogLocker.Gui.Avalonia"
+             x:Class="ColDogStudios.ColDogLocker.Avalonia.App"
+             xmlns:local="using:ColDogStudios.ColDogLocker.Avalonia"
              RequestedThemeVariant="Default">
              <!-- "Default" ThemeVariant follows system theme variant. "Dark" or "Light" are other available options. -->
 
     <Application.DataTemplates>
         <local:ViewLocator/>
     </Application.DataTemplates>
-  
+
     <Application.Styles>
         <FluentTheme />
     </Application.Styles>
