+++ ColDogLocker.Avalonia/App.axaml.cs
@@ -1,0 +1,28 @@
+using Avalonia;
+using Avalonia.Controls.ApplicationLifetimes;
+using Avalonia.Data.Core;
+using Avalonia.Data.Core.Plugins;
+using System.Linq;
+using Avalonia.Markup.Xaml;
+using ColDogStudios.ColDogLocker.Avalonia.ViewModels;
+using ColDogStudios.ColDogLocker.Avalonia.Views;
+
+namespace ColDogStudios.ColDogLocker.Avalonia;
+
+public partial class App : Application
+{
+    public override void Initialize()
+    {
+        AvaloniaXamlLoader.Load(this);
+    }
+
+    public override void OnFrameworkInitializationCompleted()
+    {
+        if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
+        {
+            desktop.MainWindow = new MainWindow { DataContext = new MainWindowViewModel(), };
+        }
+
+        base.OnFrameworkInitializationCompleted();
+    }
+}
