+++ ColDogLocker.Avalonia/Assets/cdlIcon.ico
