+++ ColDogLocker.Avalonia/ViewModels/MainWindowViewModel.cs
@@ -1,0 +1,6 @@
+﻿namespace ColDogStudios.ColDogLocker.Avalonia.ViewModels;
+
+public partial class MainWindowViewModel : ViewModelBase
+{
+    public string Greeting { get; } = "Welcome to ColDog Locker!\nThis GUI is currently under construction.";
+}
