+++ ColDogLocker.Avalonia/ViewModels/ViewModelBase.cs
@@ -1,0 +1,7 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace ColDogStudios.ColDogLocker.Avalonia.ViewModels;
+
+public abstract class ViewModelBase : ObservableObject
+{
+}
