+++ ColDogLocker.Avalonia/Views/MainWindow.axaml
@@ -1,13 +1,15 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:vm="using:ColDogLocker.Gui.Avalonia.ViewModels"
+        xmlns:vm="using:ColDogStudios.ColDogLocker.Avalonia.ViewModels"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="ColDogLocker.Gui.Avalonia.Views.MainWindow"
+        x:Class="ColDogStudios.ColDogLocker.Avalonia.Views.MainWindow"
         x:DataType="vm:MainWindowViewModel"
-        Icon="/Assets/avalonia-logo.ico"
-        Title="ColDogLocker.Gui.Avalonia">
+        Icon="/Assets/cdlIcon.ico"
+        Title="ColDog Locker"
+        CanResize="False"
+        CanMaximize="False">
 
     <Design.DataContext>
         <!-- This only sets the DataContext for the previewer in an IDE,
