--- ColDogLocker.Gui.Avalonia/ViewModels/MainWindowViewModel.cs
@@ -1,7 +1,0 @@
-﻿namespace ColDogLocker.Gui.Avalonia.ViewModels
-{
-    public partial class MainWindowViewModel : ViewModelBase
-    {
-        public string Greeting { get; } = "Welcome to Avalonia!";
-    }
-}
