--- ColDogLocker.Gui.Avalonia/ViewModels/ViewModelBase.cs
@@ -1,8 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-
-namespace ColDogLocker.Gui.Avalonia.ViewModels
-{
-    public abstract class ViewModelBase : ObservableObject
-    {
-    }
-}
