--- ColDogLocker.Gui.Avalonia/Views/MainWindow.axaml.cs
@@ -1,12 +1,0 @@
-using Avalonia.Controls;
-
-namespace ColDogLocker.Gui.Avalonia.Views
-{
-    public partial class MainWindow : Window
-    {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
-    }
-}
