+++ ColDogLocker.slnx
@@ -1,10 +1,10 @@
 <Solution>
   <Project Path="ColDogLocker.Application.Tests/ColDogLocker.Application.Tests.csproj" />
   <Project Path="ColDogLocker.Application/ColDogLocker.Application.csproj" />
+  <Project Path="ColDogLocker.Avalonia/ColDogLocker.Avalonia.csproj" />
   <Project Path="ColDogLocker.Cli/ColDogLocker.Cli.csproj" />
   <Project Path="ColDogLocker.Core.Tests/ColDogLocker.Core.Tests.csproj" />
   <Project Path="ColDogLocker.Core/ColDogLocker.Core.csproj" />
-  <Project Path="ColDogLocker.Gui.Avalonia/ColDogLocker.Gui.Avalonia.csproj" Id="cf8be687-28a9-4a5c-8bec-e198c0731b2f" />
   <Project Path="ColDogLocker.Gui.WPF/ColDogLocker.Gui.WPF.csproj" Id="16cced71-e057-4a5e-9649-b580f104945c" />
   <Project Path="ColDogLocker.Gui/ColDogLocker.Gui.csproj" />
   <Project Path="ColDogLocker.Infrastructure.Tests/ColDogLocker.Infrastructure.Tests.csproj" />
